fix: move some util where they are more reusable
