+++ packages/inter-protocol/test/stakeFactory/test-stakeFactory.js
@@ -5,9 +5,8 @@ import { AmountMath, AssetKind, makeIssuerKit } from '@agoric/ertp';
 import binaryVoteCounterBundle from '@agoric/governance/bundles/bundle-binaryVoteCounter.js';
 import committeeBundle from '@agoric/governance/bundles/bundle-committee.js';
 import contractGovernorBundle from '@agoric/governance/bundles/bundle-contractGovernor.js';
-import { makeCopyBag } from '@agoric/store';
+import { makeCopyBag, objectMap } from '@agoric/store';
 import { unsafeMakeBundleCache } from '@agoric/swingset-vat/tools/bundleTool.js';
-import { objectMap } from '@agoric/vat-data';
 import centralSupplyBundle from '@agoric/vats/bundles/bundle-centralSupply.js';
 import mintHolderBundle from '@agoric/vats/bundles/bundle-mintHolder.js';
 import {
