+++ packages/vat-data/src/kind-utils.js
@@ -1,4 +1,4 @@
-import { provide } from '@agoric/store';
+import { provide, objectMap } from '@agoric/store';
 import {
   defineDurableKind,
   defineDurableKindMulti,
@@ -11,8 +11,6 @@ import {
 /** @template T @typedef {import('./types.js').KindFacet<T>} KindFacet */
 /** @template T @typedef {import('./types.js').KindFacets<T>} KindFacets */
 /** @typedef {import('./types.js').DurableKindHandle} DurableKindHandle */
-
-const { entries, fromEntries } = Object;
 
 /**
  * Make a version of the argument function that takes a kind context but ignores it.
@@ -35,53 +33,6 @@ export const provideKindHandle = (baggage, kindName) =>
   provide(baggage, `${kindName}_kindHandle`, () => makeKindHandle(kindName));
 // @ts-expect-error TODO statically recognize harden
 harden(provideKindHandle);
-
-/**
- * By analogy with how `Array.prototype.map` will map the elements of
- * an array to transformed elements of an array of the same shape,
- * `objectMap` will do likewise for the string-named own enumerable
- * properties of an object.
- *
- * Typical usage applies `objectMap` to a CopyRecord, i.e.,
- * an object for which `passStyleOf(original) === 'copyRecord'`. For these,
- * none of the following edge cases arise. The result will be a CopyRecord
- * with exactly the same property names, whose values are the mapped form of
- * the original's values.
- *
- * When the original is not a CopyRecord, some edge cases to be aware of
- *    * No matter how mutable the original object, the returned object is
- *      hardened.
- *    * Only the string-named enumerable own properties of the original
- *      are mapped. All other properties are ignored.
- *    * If any of the original properties were accessors, `Object.entries`
- *      will cause its `getter` to be called and will use the resulting
- *      value.
- *    * No matter whether the original property was an accessor, writable,
- *      or configurable, all the properties of the returned object will be
- *      non-writable, non-configurable, data properties.
- *    * No matter what the original object may have inherited from, and
- *      no matter whether it was a special kind of object such as an array,
- *      the returned object will always be a plain object inheriting directly
- *      from `Object.prototype` and whose state is only these new mapped
- *      own properties.
- *
- * With these differences, even if the original object was not a CopyRecord,
- * if all the mapped values are Passable, then the returned object will be
- * a CopyRecord.
- *
- * @template {string} K
- * @template T
- * @template U
- * @param {Record<K,T>} original
- * @param {(value: T, key?: string) => U} mapFn
- * @returns {Record<K,U>}
- */
-export const objectMap = (original, mapFn) => {
-  const ents = entries(original);
-  const mapEnts = ents.map(([k, v]) => [k, mapFn(v, k)]);
-  // @ts-expect-error TODO statically recognize harden
-  return /** @type {Record<K, U>} */ (harden(fromEntries(mapEnts)));
-};
 
 /** @type {import('./types.js').VivifyKind} */
 export const vivifyKind = (
